refactor in-memory db test setup to a helper file
