+++ tests/app.test.js
@@ -2,32 +2,8 @@ const app = require('../app');
 const GJV = require('geojson-validation');
 const request = require('supertest');
 const api = request(app);
-const mongoose = require('mongoose');
-const {MongoMemoryServer} = require('mongodb-memory-server');
-const Bin = require('../models/bin');
-const fs = require('fs');
-const contents = fs.readFileSync(__dirname + '/testData.json');
-const jsonContent = JSON.parse(contents);
+require('./inMemoryDatabaseTestHelper')
 
-
-beforeAll(async () => {
-  mongoServer = new MongoMemoryServer();
-  const mongoUri = await mongoServer.getConnectionString();
-  await mongoose.connect(mongoUri,  { useNewUrlParser: true });
-});
-
-afterAll(async () => {
-  await new Promise((resolve) => setTimeout(() => {
-    resolve();
-  }, 500)); // avoid jest open handle error with jest and supertest
-  await mongoose.disconnect();
-  await mongoServer.stop();
-});
-
-beforeEach(async () => {
-  await Bin.deleteMany({});
-  await Bin.insertMany(jsonContent)
-});
 
 describe('main page loads', () => {
   test('shows main page', () => {
